+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vegastack/pages-core
 
+## 0.2.0-next.13
+
 ## 0.2.0-next.12
 
 ## 0.2.0-next.11